requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2